requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
django-cors-headers>=4.0,<4.1
